FIAP/Enterprise Challenge - Sprint 4/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.19
pandas==2.1.2
numpy==1.26.1
scikit-learn==1.3.0
joblib==1.3.2
matplotlib==3.8.1
paho-mqtt==1.6.1
requests==2.31.0
streamlit==1.50.0